// backend/modules/discord/discord.interfaces.ts
export interface DiscordCmdParamsMetadataApi {
    /**
     * Minimum amount of positional parameters that are required to be passed.
     * Equals to `.definition.length` without optional array param at the end
     * by default.
     */
    minRequiredAmount?: Nullable<number>;

    /**
     * Maximum amount of positional parameters that may be passed to the command.
     * By default equals to `.definition.length` if there is no rest parameters
     * defined (`Joi.ArraySchema` at the end of `.definition`), `65535` otherwise.
     * 
     * Pre: `max >= minRequired`
     */
    maxAmount?: Nullable<number>;
    
    /**
     * Defines the tuple of arguments that the given command handler accepts.
     * Pre: `.definition.length > 0`.
     */
    definition: DiscordCmdParamMetadataApi[];
}

export interface DiscordCmdParamMetadataApi {
    /**
     * Descriptive and short name of the parameter to display in help message.
     */
    readonly name: string;

    /**
     * Short explanation what this paramter is used for.
     */
    readonly description: string;

    /** 
     * Defines type limitations of the parameter, if it is anything other than
     * `Joi.StringSchema`, it gets transformed to the expected type before passing it
     * to the command handler.
     * If schema is not specified, no type checks are performed and paramter gets
     * forwarded as a simple string to the handler. However parameter is considered to
     * be required if only if it is not positioned after other not required parameters.
     * 
     * Pre: If schema is not `.required()` all parameters that appear after this
     * one must not be `.required()` too.
     * 
     * Pre: If it is `Joi.ArraySchema` it must be the last paramter in the param list.
     */
    readonly schema?: Nullable<DiscordCmdParamSchema>;
}
export type DiscordCmdScalarParam = boolean | number | string;

export type DiscordCmdParamSchema = 
    | Joi.StringSchema 
    | Joi.NumberSchema 
    | Joi.BooleanSchema
    | Joi.ArraySchema<Joi.Value<DiscordCmdScalarParam[]>>;